+++ CHANGELOG.md
@@ -1,5 +1,8 @@
-# dev
+# 1.1.6 2020-04-11
 
+- mask JSON works with cached GDS files for the klayout placer
+- added layers in the config
+- write_labels get layer from `CONFIG['layers']['LABEL']`
 - add_padding works over the same component
 - import_gds can snap points to a design grid